apiserver/audit: split apart legacy funcs
